refactor(voice/capture): extract pure helpers into capture subpackage [T1.4 step 4]
Master mission Phase 1 / T1.4 — fourth stage. Builds on steps 1-3
(`32efb40`, `c960f6c`, `3901dfb`) by extracting the remaining
pure helpers — RMS dBFS, dBFS regex parsing, device-entry resolver
— so the only thing left in ``_capture_task.py`` (besides the
re-export shim) is the ``AudioCaptureTask`` class itself + the
module-level constants tied to its threading invariants.

What lands:

  * New module ``capture/_helpers.py`` containing:
      - ``_RMS_FLOOR_DB`` — silence-floor dBFS constant.
      - ``_rms_db_int16`` — dBFS RMS of an int16 buffer (numpy
        lazy-imported inside the function).
      - ``_PEAK_DB_RE`` — opener silence-attempt detail regex.
      - ``_extract_peak_db`` — parse ``peak -XX.X dBFS`` from
        a detail string.
      - ``_resolve_input_entry`` — selector → DeviceEntry
        resolver (int / str / None, with optional host_api
        refinement).
    Total ~155 LOC. ``DeviceEntry`` + ``Callable`` typed-only
    via ``TYPE_CHECKING``.
  * ``_capture_task.py`` drops the five definitions, adds
    ``import X as X`` re-exports for each name. ``test_capture_task.py``
    imports ``_extract_peak_db`` and ``_resolve_input_entry``
    directly from the legacy path; the re-export keeps that
    test working unchanged.
  * ``capture/__init__.py`` re-exports the five names alongside
    the steps 1-3 surfaces.
  * Removed unused ``import re`` from ``_capture_task.py`` (no
    other ``re`` usage post-extraction; auto-fix by ruff).

T1.4 cumulative impact (steps 1-4):

  * ``_capture_task.py``: 2785 → 2228 LOC (~557 LOC extracted,
    ~20 % of the split done).
  * Four coherent sub-modules under ``voice/capture/``:
    ``_restart.py`` (verdict types), ``_exceptions.py`` (error
    hierarchy), ``_contention.py`` (session-manager classifier),
    ``_helpers.py`` (RMS / dBFS / device resolver). All
    pure-types / pure-functions content with zero class state
    coupling and zero behaviour change.
  * Remaining: the ``AudioCaptureTask`` class itself (~1972 LOC)
    + module-level constants tightly tied to its threading
    invariants (sample rate, frame size, ring-buffer epoch
    packing, sustained-underrun detection state). The
    ``AudioCaptureTask`` class via the ``_RestartMixin`` pattern
    is the genuinely-risky "DEDICATED SESSION" surgery the spec
    flagged; future steps can land independently with the legacy
    stub intact.

Validation:
  * ruff check + format clean (auto-fix removed unused
    ``import re``)
  * mypy strict: 0 issues across 423 source files (+1 source file
    for ``capture/_helpers.py``)
  * bandit clean
  * pytest full suite: 11894 passed, 25 skipped, 57 deselected
    (zero delta — pure structural extraction, no test changes
    needed because the legacy import surface is preserved
    verbatim via ``import X as X`` re-exports)

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 (step 4)
